Implement new PICO-8 0.2.0 behaviour for negative division overflows.
This bug was discussed here: https://www.lexaloffle.com/bbs/?tid=36692
however the 0.2.0 changelog does not mention this fix. It also mentions
a “Fixed: "-32768 % y" gives wrong results” entry for which I could not
find a bug report. Not sure whether the two bugs are related and/or if
we’re affected by this second bug.